pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "visual-storytelling-monorepo"
version = "0.1.0"
description = "Documentation-first monorepo for grounded multi-image storytelling."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "gradio>=5.0",
  "pydantic>=2.7",
  "python-dotenv>=1.0",
  "Pillow>=10.0",
  "openai>=1.0",
  "instructor>=1.0",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.10",
  "pytest>=8.2",
  "pytest-cov>=5.0",
  "ruff>=0.5",
  "selenium>=4.41",
]

[tool.setuptools]

[tool.setuptools.packages.find]
where = ["."]
include = ["packages*"]

[tool.pytest.ini_options]
addopts = "-ra"
pythonpath = [
  ".",
  "apps/api/src",
  "packages/core/src",
  "packages/services/src",
  "packages/infra/src",
]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
src = ["apps", "packages", "tests"]
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
namespace_packages = true
explicit_package_bases = true
mypy_path = [
  ".",
  "apps/api/src",
  "packages/core/src",
  "packages/services/src",
  "packages/infra/src",
]
